Refactors deleteFood function
