+++ lib/ajax-requests/deleteFood.js
@@ -1,3 +1,20 @@
 const $ = require('jquery')
-const API = 'https://api-qs.herokuapp.com'
+const requests = require('../ajax-responses/foodResponses')
 
+function deleteFood () {
+  const foodId = getId(this)
+  return $.ajax({
+    url: `https://api-qs.herokuapp.com/api/v1/foods/${foodId}`,
+    method: 'DELETE',
+  })
+  .then(function() {
+    $(`.food${foodId}`).remove()
+  })
+  .catch(requests.errorLog) 
+}
+
+function getId (id) {
+  return id.className.split(' ')[1]
+}
+
+module.exports = {deleteFood}
